chroma_project_demo/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
chromadb==0.4.18
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-core>=0.1.20
openai>=1.10.0
requests==2.31.0
pydantic>=2.5.0
google-generativeai==0.3.2
huggingface-hub>=0.20.0
sentence-transformers==2.2.2
# OCR & image/PDF helpers
Pillow==10.4.0
pytesseract==0.3.13
pypdfium2==4.30.0
pdfplumber==0.11.4
opencv-python-headless==4.10.0.84
chardet==5.2.0
camelot-py==0.11.0
ghostscript==0.8.1
PyMuPDF==1.24.9


# Excel/CSV support
pandas==2.2.2
openpyxl==3.1.2
xlrd==2.0.1


# Reranking


duckduckgo-search==5.3.1b1

# Caching

# Optional S3 storage
boto3==1.34.106

redis==5.0.4
rq==1.16.1
